+++ src/tool/hpcrun/cct/cct.h
@@ -175,9 +175,9 @@ extern cct_node_t* hpcrun_cct_insert_addr(cct_node_t* cct, cct_addr_t* addr);
 // be eliminated before writing out the cct.
 extern cct_node_t* hpcrun_cct_insert_dummy(cct_node_t* node, uint16_t lm_ip);
 
-extern cct_node_t* hpcrun_cct_insert_range(cct_node_t* node, uint16_t lm_ip);
+extern cct_node_t* hpcrun_cct_insert_range(cct_node_t* node, uintptr_t lm_ip);
 
-extern cct_node_t* hpcrun_cct_insert_context(cct_node_t* node, uint16_t lm_ip);
+extern cct_node_t* hpcrun_cct_insert_context(cct_node_t* node, uintptr_t lm_ip);
 
 //
 // 2nd fundamental mutator: mark a node as "terminal". That is,
